In setRect(), don't return a pointer to stack memory
Noticed a compiler warning for this during build, and the code looks
like the sort of thing that may often work accidentally, but is not
at all guaranteed to work.

This is the simplest fix I could imagine; keep the logic about whether
to return NULL or not in setRect(), but pass it a reference to some
memory in the caller's stack frame to use for the result.

This fixes the warning, but I haven't had a chance to verify functionality.