Add GetStructuringElement function
GetStructuringElement returns a kernel that can be used to perform
morphological operation. This function can be available in `cv.imgproc`
and takes two arguments, the erosion type (MORPH_RECT, MORPH_CROSS,
MORPH_ELLIPSE, ...) and the size of the generated kernel. The `ksize`
argument must be an array of two numbers.
The output kernel is often use for Dilation and Erosion operations.